+++ src/videodownlinkbandwidthpolicy/AllHighestVideoBandwidthPolicy.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import ClientMetricReport from '../clientmetricreport/ClientMetricReport';
+import VideoSource from '../videosource/VideoSource';
 import DefaultVideoStreamIdSet from '../videostreamidset/DefaultVideoStreamIdSet';
 import VideoStreamIdSet from '../videostreamidset/VideoStreamIdSet';
 import VideoStreamIndex from '../videostreamindex/VideoStreamIndex';
@@ -15,6 +16,9 @@ import VideoDownlinkBandwidthPolicy from './VideoDownlinkBandwidthPolicy';
 export default class AllHighestVideoBandwidthPolicy implements VideoDownlinkBandwidthPolicy {
   private optimalReceiveSet: VideoStreamIdSet;
   private subscribedReceiveSet: VideoStreamIdSet;
+  private areVideoSourcesChosen: boolean;
+  private videoSources: VideoSource[];
+  protected videoIndex: VideoStreamIndex;
 
   constructor(private selfAttendeeId: string) {
     this.reset();
@@ -23,10 +27,17 @@ export default class AllHighestVideoBandwidthPolicy implements VideoDownlinkBand
   reset(): void {
     this.optimalReceiveSet = new DefaultVideoStreamIdSet();
     this.subscribedReceiveSet = new DefaultVideoStreamIdSet();
+    this.videoSources = [];
+    this.areVideoSourcesChosen = false;
   }
 
   updateIndex(videoIndex: VideoStreamIndex): void {
-    this.optimalReceiveSet = this.calculateOptimalReceiveSet(videoIndex);
+    this.videoIndex = videoIndex;
+    if (!this.areVideoSourcesChosen) {
+      this.optimalReceiveSet = this.calculateOptimalReceiveSet(videoIndex);
+    } else {
+      this.chooseRemoteVideoSources(this.videoSources);
+    }
   }
 
   updateMetrics(_clientMetricReport: ClientMetricReport): void {}
@@ -38,6 +49,39 @@ export default class AllHighestVideoBandwidthPolicy implements VideoDownlinkBand
   chooseSubscriptions(): VideoStreamIdSet {
     this.subscribedReceiveSet = this.optimalReceiveSet.clone();
     return this.subscribedReceiveSet.clone();
+  }
+
+  chooseRemoteVideoSources(videoSources: VideoSource[]): void {
+    if (!this.videoIndex || this.videoIndex.allStreams().empty()) {
+      return;
+    }
+
+    this.videoSources = videoSources;
+
+    // Get the list of all the remote stream information
+    const remoteInfos = this.videoIndex.remoteStreamDescriptions();
+
+    // calculate the optimal set for the given index
+    const receiveSet = this.calculateOptimalReceiveSet(this.videoIndex);
+
+    const mapOfAttendeeIdToOptimalStreamId = new Map<string, number>();
+
+    for (const info of remoteInfos) {
+      if (receiveSet.contain(info.streamId)) {
+        mapOfAttendeeIdToOptimalStreamId.set(info.attendeeId, info.streamId);
+      }
+    }
+
+    const streamSelectionSet = new DefaultVideoStreamIdSet();
+    for (const videoSource of videoSources) {
+      const attendeeId = videoSource.attendee.attendeeId;
+      if (mapOfAttendeeIdToOptimalStreamId.has(attendeeId)) {
+        streamSelectionSet.add(mapOfAttendeeIdToOptimalStreamId.get(attendeeId));
+      }
+    }
+
+    this.areVideoSourcesChosen = true;
+    this.optimalReceiveSet = streamSelectionSet.clone();
   }
 
   private calculateOptimalReceiveSet(videoIndex: VideoStreamIndex): VideoStreamIdSet {