+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Ability to choose remote video sources in `AllHighestVideoBandwidthPolicy`
+
+### Removed
+
+### Fixed
+
+### Changed
+- Enforced a video receive limit incase the number of videos shared in the meeting
+are greater 
+
 ## [2.24.0] - 2021-12-07
 ### Added
 - Add error name for custom device controller error.